.xraw parser now stores material rgba data as floats (0-1) instead of uint8s. Can now use in OpenGL land.

+++ include/xraw_parser.hpp
@@ -8,7 +8,7 @@
 
 namespace Voxel {
 	struct Material {
-		uint8_t r, g, b, a;
+		float r, g, b, a;
 
 		Material() = default;
 		Material(uint8_t r, uint8_t g, uint8_t b);
